% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linG3DClone.R
\name{linG3DClone}
\alias{linG3DClone}
\title{3D lineage tree visualization tool}
\usage{
linG3DClone(
  pathData,
  cloneNum,
  IsGradient,
  xmin,
  xmax,
  ymin,
  ymax,
  tmin,
  tmax,
  fileStep,
  toPrint
)
}
\arguments{
\item{pathData}{-- directory with input data}

\item{cloneNum}{-- clone number to be drawn}

\item{IsGradient}{-- draw drug gradient in the background 1-yes; 0-no;}

\item{xmin, xmax, ymin, ymax}{-- dimensions of the spatial domain}

\item{tmin, tmax}{-- dimensions of the temporal domain}

\item{fileStep}{-- frequency of the sampled data}

\item{toPrint}{-- save the final figure 1-yes; 0-no;}
}
\description{
This is a companion code for the paper "LinG3D: Visualizing the
Spatio-Temporal Dynamics of Clonal Evolution".
}
\details{
This code generates the 3D lineage tree of all cells from one
clone of number specified in "cloneNum" with data from the
directory "pathData".

It requires the following data in the pathData/data/ directory:.
\itemize{
\item cell_history.txt -- file with info about each cell.
\item cellID_##.txt    -- cell IDs in a file with index number.
\item cellXY_##.txt    -- cell coordinates in a file with index.
\item drug.txt         -- concentration of a drug for background.
 }
}
\examples{
for the examples discussed in the paper use:
  EXAMPLE 1: select cloneNum between 0 and 9
  linG3DClone(pathData="exampleB05", cloneNum=9,  IsGradient=1,
              xmin=-100, xmax=100, ymin=-100, ymax=100,
              tmin=0, tmax=100000, fileStep=2000, toPrint=0)

  EXAMPLE 2: select cloneNum between 0 and 147
  linG3DClone(pathData="exampleB005", cloneNum=0,  IsGradient=1,
              xmin=-100, xmax=100, ymin=-100, ymax=100,
              tmin=0, tmax=100000, fileStep=2000, toPrint=0)

}
\author{
A. Hu, A.M.E. Ojwang", K.D. Olumoyin, and K.A. Rejniak.
}
